address code reviews

+++ valhalla/mjolnir/landmark_builder.h
@@ -60,7 +60,7 @@ enum class LandmarkType : uint8_t {
   casino = 18,
 };
 
-using Landmark = std::tuple<uint32_t, std::string, LandmarkType, double, double>;
+using Landmark = std::tuple<int64_t, std::string, LandmarkType, double, double>;
 
 inline LandmarkType string_to_landmark_type(const std::string& s) {
   static const std::unordered_map<std::string, LandmarkType> string_to_landmark_type =
@@ -104,7 +104,7 @@ public:
                                                       const double maxLat,
                                                       const double maxLong);
 
-  Landmark get_landmark(const uint32_t pkey);
+  Landmark get_landmark(const int64_t pkey);
 
 protected:
   struct db_pimpl;
